Fix all warnings from the compiler
This change simply fixes all the warnings from the latest version of the
compiler. It also removes the (non-test code) unsafe blocks. One of the
warnings I think was actually a bug - a cgroup builder could fail to
apply resource constraints but still return a Cgroup silently. Now a
builder will return an Error correctly in this case - this is an API
change though so requires a mv bump.